Functional color config for NB in JSON and XML
