+++ config/SyntaxColors/Reference.txt
@@ -1,0 +1,11 @@
+JSON <-> XML converter
+http://www.utilities-online.info/xmltojson/#.VVDT9nUVhBc
+
+JSON validator
+http://jsonlint.com/
+
+XML validator
+http://www.xmlvalidation.com/index.php?id=1&L=0
+
+Color conversions
+http://www.w3schools.com/tags/ref_colorpicker.asp
