+++ nbproject/configurations.xml
@@ -8,7 +8,7 @@
         <itemPath>src/Build.h</itemPath>
         <itemPath>src/CentralGui.h</itemPath>
         <itemPath>src/CentralTabs.h</itemPath>
-        <itemPath>src/File.h</itemPath>
+        <itemPath>src/FileGui.h</itemPath>
         <itemPath>src/FileTreeGui.h</itemPath>
         <itemPath>src/GenRosPkg.h</itemPath>
         <itemPath>src/Highlighter.h</itemPath>
@@ -42,7 +42,7 @@
         <itemPath>src/Build.cpp</itemPath>
         <itemPath>src/CentralGui.cpp</itemPath>
         <itemPath>src/CentralTabs.cpp</itemPath>
-        <itemPath>src/File.cpp</itemPath>
+        <itemPath>src/FileGui.cpp</itemPath>
         <itemPath>src/FileTreeGui.cpp</itemPath>
         <itemPath>src/GenRosPkg.cpp</itemPath>
         <itemPath>src/Highlighter.cpp</itemPath>
@@ -109,9 +109,9 @@
       </item>
       <item path="src/CentralTabs.h" ex="false" tool="3" flavor2="0">
       </item>
-      <item path="src/File.cpp" ex="false" tool="1" flavor2="0">
+      <item path="src/FileGui.cpp" ex="false" tool="1" flavor2="0">
       </item>
-      <item path="src/File.h" ex="false" tool="3" flavor2="0">
+      <item path="src/FileGui.h" ex="false" tool="3" flavor2="0">
       </item>
       <item path="src/FileTreeGui.cpp" ex="false" tool="1" flavor2="0">
       </item>
@@ -222,9 +222,9 @@
       </item>
       <item path="src/CentralTabs.h" ex="false" tool="3" flavor2="0">
       </item>
-      <item path="src/File.cpp" ex="false" tool="1" flavor2="0">
+      <item path="src/FileGui.cpp" ex="false" tool="1" flavor2="0">
       </item>
-      <item path="src/File.h" ex="false" tool="3" flavor2="0">
+      <item path="src/FileGui.h" ex="false" tool="3" flavor2="0">
       </item>
       <item path="src/FileTreeGui.cpp" ex="false" tool="1" flavor2="0">
       </item>
