+++ nbproject/qt-Release.pro
@@ -8,8 +8,8 @@ CONFIG -= debug_and_release app_bundle lib_bundle
 CONFIG += release 
 PKGCONFIG +=
 QT = core gui widgets
-SOURCES += src/Build.cpp src/CentralGui.cpp src/CentralTabs.cpp src/File.cpp src/FileTreeGui.cpp src/GenRosPkg.cpp src/Highlighter.cpp src/LineNumberArea.cpp src/MasterActions.cpp src/MasterGui.cpp src/MasterMenus.cpp src/MasterToolBars.cpp src/NewFileGui.cpp src/NewProjectGui.cpp src/NewProjectPage_1.cpp src/NewProjectPage_2.cpp src/NewProjectPage_3.cpp src/OpeningGui.cpp src/OutputGui.cpp src/ScoutGui.cpp src/Terminal.cpp src/main.cpp
-HEADERS += src/Build.h src/CentralGui.h src/CentralTabs.h src/File.h src/FileTreeGui.h src/GenRosPkg.h src/Highlighter.h src/LineNumberArea.h src/MasterActions.h src/MasterGui.h src/MasterMenus.h src/MasterToolBars.h src/NewFileGui.h src/NewProjectGui.h src/NewProjectPage_1.h src/NewProjectPage_2.h src/NewProjectPage_3.h src/OpeningGui.h src/OutputGui.h src/ScoutGui.h src/Terminal.h
+SOURCES += src/Build.cpp src/CentralGui.cpp src/CentralTabs.cpp src/FileGui.cpp src/FileTreeGui.cpp src/GenRosPkg.cpp src/Highlighter.cpp src/LineNumberArea.cpp src/MasterActions.cpp src/MasterGui.cpp src/MasterMenus.cpp src/MasterToolBars.cpp src/NewFileGui.cpp src/NewProjectGui.cpp src/NewProjectPage_1.cpp src/NewProjectPage_2.cpp src/NewProjectPage_3.cpp src/OpeningGui.cpp src/OutputGui.cpp src/ScoutGui.cpp src/Terminal.cpp src/main.cpp
+HEADERS += src/Build.h src/CentralGui.h src/CentralTabs.h src/FileGui.h src/FileTreeGui.h src/GenRosPkg.h src/Highlighter.h src/LineNumberArea.h src/MasterActions.h src/MasterGui.h src/MasterMenus.h src/MasterToolBars.h src/NewFileGui.h src/NewProjectGui.h src/NewProjectPage_1.h src/NewProjectPage_2.h src/NewProjectPage_3.h src/OpeningGui.h src/OutputGui.h src/ScoutGui.h src/Terminal.h
 FORMS +=
 RESOURCES +=
 TRANSLATIONS +=
