+++ src/moc/moc_FileGui.cpp
@@ -1,0 +1,98 @@
+/****************************************************************************
+** Meta object code from reading C++ file 'FileGui.h'
+**
+** Created by: The Qt Meta Object Compiler version 63 (Qt 4.8.6)
+**
+** WARNING! All changes made in this file will be lost!
+*****************************************************************************/
+
+#include "../FileGui.h"
+#if !defined(Q_MOC_OUTPUT_REVISION)
+#error "The header file 'FileGui.h' doesn't include <QObject>."
+#elif Q_MOC_OUTPUT_REVISION != 63
+#error "This file was generated using the moc from 4.8.6. It"
+#error "cannot be used with the include files from this version of Qt."
+#error "(The moc has changed too much.)"
+#endif
+
+QT_BEGIN_MOC_NAMESPACE
+static const uint qt_meta_data_File[] = {
+
+ // content:
+       6,       // revision
+       0,       // classname
+       0,    0, // classinfo
+       3,   14, // methods
+       0,    0, // properties
+       0,    0, // enums/sets
+       0,    0, // constructors
+       0,       // flags
+       0,       // signalCount
+
+ // slots: signature, parameters, type, tag, flags
+      20,    6,    5,    5, 0x08,
+      51,    5,    5,    5, 0x08,
+      76,   74,    5,    5, 0x08,
+
+       0        // eod
+};
+
+static const char qt_meta_stringdata_File[] = {
+    "File\0\0newBlockCount\0updateLineNumberAreaWidth(int)\0"
+    "highlightCurrentLine()\0,\0"
+    "updateLineNumberArea(QRect,int)\0"
+};
+
+void File::qt_static_metacall(QObject *_o, QMetaObject::Call _c, int _id, void **_a)
+{
+    if (_c == QMetaObject::InvokeMetaMethod) {
+        Q_ASSERT(staticMetaObject.cast(_o));
+        File *_t = static_cast<File *>(_o);
+        switch (_id) {
+        case 0: _t->updateLineNumberAreaWidth((*reinterpret_cast< int(*)>(_a[1]))); break;
+        case 1: _t->highlightCurrentLine(); break;
+        case 2: _t->updateLineNumberArea((*reinterpret_cast< const QRect(*)>(_a[1])),(*reinterpret_cast< int(*)>(_a[2]))); break;
+        default: ;
+        }
+    }
+}
+
+const QMetaObjectExtraData File::staticMetaObjectExtraData = {
+    0,  qt_static_metacall 
+};
+
+const QMetaObject File::staticMetaObject = {
+    { &QPlainTextEdit::staticMetaObject, qt_meta_stringdata_File,
+      qt_meta_data_File, &staticMetaObjectExtraData }
+};
+
+#ifdef Q_NO_DATA_RELOCATION
+const QMetaObject &File::getStaticMetaObject() { return staticMetaObject; }
+#endif //Q_NO_DATA_RELOCATION
+
+const QMetaObject *File::metaObject() const
+{
+    return QObject::d_ptr->metaObject ? QObject::d_ptr->metaObject : &staticMetaObject;
+}
+
+void *File::qt_metacast(const char *_clname)
+{
+    if (!_clname) return 0;
+    if (!strcmp(_clname, qt_meta_stringdata_File))
+        return static_cast<void*>(const_cast< File*>(this));
+    return QPlainTextEdit::qt_metacast(_clname);
+}
+
+int File::qt_metacall(QMetaObject::Call _c, int _id, void **_a)
+{
+    _id = QPlainTextEdit::qt_metacall(_c, _id, _a);
+    if (_id < 0)
+        return _id;
+    if (_c == QMetaObject::InvokeMetaMethod) {
+        if (_id < 3)
+            qt_static_metacall(this, _c, _id, _a);
+        _id -= 3;
+    }
+    return _id;
+}
+QT_END_MOC_NAMESPACE
